refactor: KST 로직을 별도의 공통 파일로 추출
