+++ src/backend/cache/word.cache.ts
@@ -2,19 +2,7 @@ import 'server-only';
 
 import { unstable_cache as cache } from 'next/cache';
 import { wordnikClient } from '../clients/wordnik/wordnik.clients';
-
-function getSecondsUntilKSTMidnight(): number {
-  const now = new Date();
-  const kstOffset = 9 * 60 * 60 * 1000;
-
-  const nowKST = new Date(now.getTime() + kstOffset);
-  const midnightKST = new Date(nowKST);
-
-  midnightKST.setUTCHours(24, 0, 0, 0);
-  const remainingMilliseconds = midnightKST.getTime() - nowKST.getTime();
-
-  return Math.floor(remainingMilliseconds / 1000);
-}
+import { getSecondsUntilKSTMidnight } from '@/app/utils/times';
 
 export const getCachedRandomWord = cache(
   // 실제 데이터를 가져오는 함수
